requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
dnspython==2.7.0
fastapi==0.115.6
feedparser==6.0.11
filelock==3.16.1
frozendict==2.4.6
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multitasking==0.0.11
mysql-connector-python==9.1.0
narwhals==1.18.3
numpy==2.2.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parsel==1.9.1
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
Protego==0.3.1
proto-plus==1.25.0
protobuf==5.29.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
PyDispatcher==2.0.7
Pygments==2.18.0
pymongo==4.10.1
pyOpenSSL==24.3.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
queuelib==1.7.0
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
schedule==1.2.2
Scrapy==2.6.3
scrapy-splash==0.9.0
selenium==4.27.1
service-identity==24.2.0
setuptools==75.6.0
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.41.3
streamlit==1.41.1
tenacity==9.0.0
tldextract==5.1.3
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
Twisted==22.8.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
w3lib==2.2.1
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
yfinance==0.2.50
zope.interface==7.2
